Made sure that only logged-in users can create objects and only owners can update & delete them.

+++ igdb/interaction/views.py
@@ -1,3 +1,4 @@
+from django.contrib.auth.mixins import LoginRequiredMixin
 from django.contrib.contenttypes.models import ContentType
 from django.core.exceptions import ObjectDoesNotExist
 from django.shortcuts import render, redirect, get_object_or_404
@@ -9,7 +10,7 @@ from django.views.decorators.http import require_POST
 from igdb.games.models import VideoGame, NonVideoGame
 from igdb.games.views import ReadGameView
 from igdb.interaction.forms import CreateCuratedListForm, CreateReviewForm, UpdateReviewForm, UpdateCuratedListForm, \
-    CreateLikeForm, CreateCommentForm, CreateRatingForm, UpdateCommentForm, UpdateRatingForm
+    CreateLikeForm, CreateCommentForm, CreateRatingForm, UpdateCommentForm, UpdateRatingForm, DeleteLikeForm
 from igdb.interaction.models import Review, CuratedList, Like, Comment, Rating
 
 
@@ -31,10 +32,11 @@ class ReviewsView(View):
         return render(request, context=context, template_name="reviews.html")
 
 
-class CreateCuratedListView(CreateView):
+class CreateCuratedListView(LoginRequiredMixin, CreateView):
     model = CuratedList
     form_class = CreateCuratedListForm
     template_name = "interaction\\create_list.html"
+    login_url = reverse_lazy("sign_in")
 
     def get_form_kwargs(self):
         kwargs = super().get_form_kwargs()
@@ -57,15 +59,24 @@ class ReadCuratedListView(View):
         return render(request, "interaction\\read_list.html", context)
 
 
-class UpdateCuratedListView(View):
+class UpdateCuratedListView(LoginRequiredMixin, View):
+    login_url = reverse_lazy("sign_in")
+
     def get(self, request, slug):
         curated_list = get_object_or_404(CuratedList, slug=slug)
         form = UpdateCuratedListForm(instance=curated_list)
+
+        if request.user != curated_list.user:
+            return redirect("read_list", slug=slug)
+
         return render(request, template_name="interaction\\update_list.html", context={"form": form})
 
     def post(self, request, slug):
         curated_list = get_object_or_404(CuratedList, slug=slug)
         form = CreateCuratedListForm(request.POST or None, instance=curated_list)
+
+        if request.user != curated_list.user:
+            return redirect("read_list", slug=slug)
 
         if request.POST.get("action") == "delete":
             curated_list.delete()
@@ -77,10 +88,11 @@ class UpdateCuratedListView(View):
         return render(request, template_name="interaction\\update_list.html", context={"form": form})
 
 
-class CreateReviewView(CreateView):
+class CreateReviewView(LoginRequiredMixin, CreateView):
     model = Review
     form_class = CreateReviewForm
     template_name = "interaction\\create_review.html"
+    login_url = reverse_lazy("sign_in")
 
     def get_object(self):
         slug = self.kwargs.get("slug")
@@ -116,15 +128,22 @@ class ReadReviewView(View):
         return render(request, template_name="interaction\\read_review.html", context={"review": review})
 
 
-class UpdateReviewView(View):
+class UpdateReviewView(LoginRequiredMixin, View):
+    login_url = reverse_lazy("sign_in")
+
     def get(self, request, slug):
         review = get_object_or_404(Review, slug=slug)
         form = UpdateReviewForm(instance=review)
+        if request.user != review.user:
+            return redirect("read_review", slug=slug)
         return render(request, template_name="interaction\\update_review.html", context={"form": form})
 
     def post(self, request, slug):
         review = get_object_or_404(Review, slug=slug)
         form = CreateReviewForm(request.POST or None, instance=review)
+
+        if request.user != review.user:
+            return redirect("read_review", slug=slug)
 
         if request.POST.get("action") == "delete":
             review.delete()
@@ -137,7 +156,9 @@ class UpdateReviewView(View):
         return render(request, template_name="interaction\\update_review.html", context={"form": form})
 
 
-class CreateLikeView(View):
+class CreateLikeView(LoginRequiredMixin, View):
+    login_url = reverse_lazy("sign_in")
+
     def post(self, request, slug):
         try:
             game = VideoGame.objects.get(slug=slug)
@@ -153,10 +174,26 @@ class CreateLikeView(View):
         return redirect("read_game", slug=game.slug)
 
 
-class CreateCommentView(CreateView):
+class DeleteLikeView(LoginRequiredMixin, View):
+    def post(self, request, slug):
+        try:
+            game = VideoGame.objects.get(slug=slug)
+        except VideoGame.DoesNotExist:
+            game = NonVideoGame.objects.get(slug=slug)
+
+        like = Like.objects.filter(user=request.user, object_id=game.id)
+
+        if like:
+            like.delete()
+
+        return redirect(reverse("read_game", kwargs={"slug": slug}))
+
+
+class CreateCommentView(LoginRequiredMixin, CreateView):
     model = Comment
     form_class = CreateCommentForm
     template_name = "interaction\\create_comment.html"
+    login_url = reverse_lazy("sign_in")
 
     def get_object(self):
         slug = self.kwargs.get("slug")
@@ -185,16 +222,25 @@ class CreateCommentView(CreateView):
         return reverse("read_game", kwargs={"slug": game.slug})
 
 
-class UpdateCommentView(View):
+class UpdateCommentView(LoginRequiredMixin, View):
+    login_url = reverse_lazy("sign_in")
+
     def get(self, request, pk):
         comment = get_object_or_404(Comment, id=pk)
         form = UpdateCommentForm(instance=comment)
+
+        if request.user != comment.user:
+            return redirect("read_game", slug=comment.content_object.slug)
+
         return render(request, template_name="interaction\\update_comment.html", context={"form": form})
 
     def post(self, request, pk):
         comment = get_object_or_404(Comment, id=pk)
         form = CreateCommentForm(request.POST or None, instance=comment)
         slug = comment.content_object.slug
+
+        if request.user != comment.user:
+            return redirect("read_game", slug=slug)
 
         if request.POST.get("action") == "delete":
             comment.delete()
@@ -207,10 +253,11 @@ class UpdateCommentView(View):
         return render(request, template_name="interaction\\update_comment.html", context={"form": form})
 
 
-class CreateRatingView(CreateView):
+class CreateRatingView(LoginRequiredMixin, CreateView):
     model = Rating
     form_class = CreateRatingForm
     template_name = "interaction\\create_rating.html"
+    login_url = reverse_lazy("sign_in")
 
     def get_object(self):
         slug = self.kwargs.get("slug")
@@ -239,16 +286,25 @@ class CreateRatingView(CreateView):
         return reverse("read_game", kwargs={"slug": game.slug})
 
 
-class UpdateRatingView(View):
+class UpdateRatingView(LoginRequiredMixin, View):
+    login_url = reverse_lazy("sign_in")
+
     def get(self, request, pk):
         rating = get_object_or_404(Rating, id=pk)
         form = UpdateRatingForm(instance=rating)
+
+        if request.user != rating.user:
+            return redirect("read_game", slug=rating.content_object.slug)
+
         return render(request, template_name="interaction\\update_rating.html", context={"form": form})
 
     def post(self, request, pk):
         rating = get_object_or_404(Rating, id=pk)
         form = CreateCommentForm(request.POST or None, instance=rating)
         slug = rating.content_object.slug
+
+        if request.user != rating.user:
+            return redirect("read_game", slug=rating.content_object.slug)
 
         if request.POST.get("action") == "delete":
             rating.delete()
